Coerce types of int and bool

+++ lib/src/auth/sasl_authenticator.dart
@@ -49,7 +49,7 @@ abstract class SaslAuthenticator extends Authenticator {
 
       currentStep = currentStep.transition(conversation, payloadAsBytes);
 
-      var conversationId = result['conversationId'];
+      var conversationId = result['conversationId'] as int;
 
       command = DbCommand.createSaslContinueCommand(db.authSourceDb ?? db,
           conversationId, currentStep.bytesToSendToServer);
